pyproject.toml
[project]
name = "disaster-triage-env"
version = "0.1.0"
description = "OpenEnv-compliant Disaster Triage RL Environment"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
inference = [
    "torch>=2.2.0",
    "gymnasium>=0.29.0",
]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
